+++ src/js/options.ts
@@ -1,8 +1,10 @@
 import type { BrowserTransportOptions } from '@sentry/browser/types/transports/types';
 import type { ProfilerProps } from '@sentry/react/types/profiler';
 import type { CaptureContext, ClientOptions, Options } from '@sentry/types';
+import { Platform } from 'react-native';
 
 import type { TouchEventBoundaryProps } from './touchevents';
+import { getExpoConstants } from './utils/expomodules';
 
 export interface BaseReactNativeOptions {
   /**
@@ -181,4 +183,29 @@ export interface ReactNativeWrapperOptions {
 
   /** Props for the root touch event boundary */
   touchEventBoundaryProps?: TouchEventBoundaryProps;
+}
+
+/**
+ * If the user has not explicitly set `enableNativeNagger`
+ * the function enables native nagging based on the current
+ * environment.
+ */
+export function shouldEnableNativeNagger(userOptions: unknown): boolean {
+  if (typeof userOptions === 'boolean') {
+    // User can override the default behavior
+    return userOptions;
+  }
+
+  if (Platform.OS === 'web' || Platform.OS === 'windows') {
+    // We don't want to nag on known platforms that don't support native
+    return false;
+  }
+
+  const expoConstants = getExpoConstants();
+  if (expoConstants && expoConstants.appOwnership === 'expo') {
+    // If the app is running in Expo Go, we don't want to nag
+    return false;
+  }
+
+  return true;
 }