+++ src/js/utils/environment.ts
@@ -1,6 +1,7 @@
 import { Platform } from 'react-native';
 
 import { RN_GLOBAL_OBJ } from '../utils/worldwide';
+import { getExpoConstants } from './expomodules';
 import { ReactNativeLibraries } from './rnlibraries';
 
 /** Checks if the React Native Hermes engine is running */
@@ -30,6 +31,26 @@ export function getReactNativeVersion(): string | undefined {
 /** Checks if Expo is present in the runtime */
 export function isExpo(): boolean {
   return RN_GLOBAL_OBJ.expo != null;
+}
+
+/** Check if JS runs in Expo Go */
+export function isExpoGo(): boolean {
+  const expoConstants = getExpoConstants();
+  return (expoConstants && expoConstants.appOwnership) === 'expo';
+}
+
+/** Check Expo Go version if available */
+export function getExpoGoVersion(): string | undefined {
+  const expoConstants = getExpoConstants();
+  return typeof expoConstants?.expoVersion === 'string' ? expoConstants.expoVersion : undefined;
+}
+
+/** Returns Expo SDK version if available */
+export function getExpoSdkVersion(): string | undefined {
+  const expoConstants = getExpoConstants();
+  const [, expoSdkVersion] =
+    typeof expoConstants?.manifest?.runtimeVersion === 'string' ? expoConstants.manifest.runtimeVersion.split(':') : [];
+  return expoSdkVersion;
 }
 
 /** Checks if the current platform is not web */